backend/requirements.txt
fastapi
uvicorn
torch
openai
python-dotenv
pydantic
